source_code/test_security_final.py
#!/usr/bin/env python3
"""
보안 기능 최종 검증 스크립트
"""

import os
import sys
sys.path.append('.')

# 환경변수 로드
from dotenv import load_dotenv
load_dotenv()

from src.api.security import SecurityConfig, RateLimitConfig, get_security_stats

def main():
    print('🔒 보안 기능 최종 검증')
    print('=' * 50)

    # 환경변수에서 보안 설정 생성
    security_config = SecurityConfig(
        max_file_size=int(os.getenv('MAX_FILE_SIZE_MB', '10')) * 1024 * 1024,
        allowed_file_types=os.getenv('ALLOWED_MIME_TYPES', 'audio/wav,audio/x-wav').split(','),
        allowed_extensions=os.getenv('ALLOWED_FILE_EXTENSIONS', '.wav').split(','),
        force_https=os.getenv('FORCE_HTTPS', 'false').lower() == 'true',
        rate_limit=RateLimitConfig(
            max_requests=int(os.getenv('RATE_LIMIT_REQUESTS', '100')),
            time_window=int(os.getenv('RATE_LIMIT_WINDOW', '3600')),
            block_duration=int(os.getenv('RATE_LIMIT_BLOCK', '3600'))
        ),
        trusted_proxies=[ip.strip() for ip in os.getenv('TRUSTED_PROXIES', '').split(',') if ip.strip()]
    )

    print('✅ 1. 파일 업로드 크기 제한')
    print(f'   최대 파일 크기: {security_config.max_file_size / (1024*1024):.1f}MB')

    print('✅ 2. 파일 형식 검증')
    print(f'   허용 확장자: {security_config.allowed_extensions}')
    print(f'   허용 MIME 타입: {security_config.allowed_file_types}')

    print('✅ 3. Rate limiting')
    print(f'   최대 요청 수: {security_config.rate_limit.max_requests}')
    print(f'   시간 창: {security_config.rate_limit.time_window}초')
    print(f'   차단 지속 시간: {security_config.rate_limit.block_duration}초')

    print('✅ 4. HTTPS 강제 설정')
    print(f'   HTTPS 강제: {security_config.force_https}')

    print('✅ 5. 신뢰할 수 있는 프록시')
    proxy_list = security_config.trusted_proxies if security_config.trusted_proxies else '없음'
    print(f'   프록시 목록: {proxy_list}')

    # 보안 통계 확인
    stats = get_security_stats()
    print()
    print('📊 보안 통계')
    print(f'   차단된 IP 수: {stats["blocked_ips"]}')
    print(f'   활성 클라이언트 수: {stats["active_clients"]}')

    print()
    print('🎉 모든 보안 기능이 정상적으로 구현되었습니다!')
    print()
    print('📋 구현된 보안 기능:')
    print('   • 파일 업로드 크기 제한 (최대 10MB)')
    print('   • 파일 형식 검증 (WAV 파일만 허용)')
    print('   • Rate limiting (요청 빈도 제한)')
    print('   • HTTPS 통신 강제 설정')
    print('   • 보안 헤더 추가')
    print('   • 파일명 보안 검증')
    print('   • 파일 내용 검증 (WAV 헤더 확인)')
    print('   • 클라이언트 IP 추출 및 프록시 지원')

if __name__ == "__main__":
    main()